requirements.txt
requests==2.22.0
bert_extractive_summarizer==0.3.0
joblib==0.14.1
spacy==2.2.3
datefinder==0.7.0
matplotlib==3.1.3
pandas==1.0.1
numpy==1.18.1
Pillow==7.0.0
beautifulsoup4==4.8.2
progressbar2==3.47.0
atlas==0.27.0
summarizer==0.0.7
streamlit==0.55.0

https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm==2.2.0